Update for core interwiki user handling change
MediaWiki core change I5401941 is doing two things that are relevant to
CentralAuth:

* The 'byText' parameter to Block may no longer name a usable name that
  doesn't exist locally.
* Imports may call a new ImportHandleUnknownUser hook to give extensions
  a chance to auto-create users.

For the first, we introduce a new $wgCentralAuthGlobalBlockInterwikiPrefix
variable that holds an interwiki prefix to prepend to the value passed
to byText to make it an invalid username. Core changes to Linker will
interpret this properly in block logs and generate an interwiki link.

For the second, we implement the hook to auto-create the local user if
the central user exists.

Note that, while this should work ok without I5401941, it'll start
producing global blocks with interwiki usernames that core will display
in a somewhat strange manner.

Bug: T9240
Bug: T111605
Change-Id: Ieae5d24f9098c1977447c50a8d4e2cab58a24d9f
Depends-On: I4e8b882597d2ed8075a71393e04bb0c36eebb47c