Update patch set 2
Patch Set 2:

Ping?

I did a full search on my entire dev environment and found about 350 calls to this StatusValue::fatal() method. About 240 are not trivial (which means they contain more than just a message key). I could not find a single one that utilizes any form or escaping. Page names, even URLs are used as parameters, but nothing is (manually) escaped when passed to StatusValue::fatal().

I might have entirely misunderstood the concern raised in the comments before. If you think this is the case, please help me understand.

All this patch aimed to do is to fix a super-obvious bug: the message *requires* the first parameter to be set. The current code is broken and shows an unresolved "$1" to the user.

Patch-set: 2
Label: Verified=0