Update patch set 4
Patch Set 4: Code-Review-1

User::getCanonicalName() has a parameter named $validate. Gergő was telling you to set that parameter to false.

T35602 was an old bug from when ApiQueryAllUsers did more validation, much like what you're adding here. Since we seldom clean up old usernames that become invalid, it's possible that such a name would be encountered while enumerating all users, and if things work out right that invalid name might be the one used for a continuation. Which blows up if validation is applied to that name.

The more I think about this, though, I think the right thing to do would be to add full validation to agufrom, but stop using agufrom as the continuation parameter. Instead you'd introduce an "agucontinue" that works the same as agufrom but is not transformed or validated. That way T180858 gets the DWIM it wants while not running into weird issues if continuation hits a name that breaks the assumptions the DWIM is making.

Patch-set: 4
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1