Update patch set 2
Patch Set 2:

This is such a bad argument to vote a patch down, I don't even know where to start. What annoys me the most is that there is no way forward. A -1 literally means "please improve", but you don't even say what you disagree with. It's not clear to which "we" you refer to. It's not clear how this affects core, and why do you think so. Your initial claim is even wrong.

Patch-set: 2