Update patch set 2
Patch Set 2:

I really don't care about this line in the .phpcs.xml file. That's not the point of this patch.

It appears you perceive this as a proposal to change something in core or the MediaWiki CodeSniffer. Neither is the case. This is the CentralAuth extension, and nothing else. Please point me to the consensus on mediawiki.org that forbids extensions to have additional PHPCS sniffs enabled.

Patch-set: 2
Label: Verified=0