LocalRenameUserJob: escape backreferences in replacement title
preg_replace treats a sequence in the form "\n", "$n", and "${n}" as
a back reference to a captured pattern in the regular expression. If no
such pattern is found it will be replaced with the empty string. This
causes usernames like "Drytime%$1600" to be changed to "Drytime%00" when
passed through preg_replace in the user page renaming loop. Escaping the
'$' or '\' with two backslashes ('\\') will prevent this behavior.

Bug: T188171
Change-Id: I82bbb2cf380f4bd43e679d416b953a3a943b27d1
(cherry picked from commit 3f0f002168091fca3e45869bf1165f764115a2bb)