Minor escaping fixes for phan-taint-check
This does not fix all the phan-taint-check issues, just
some of them. The fix for migrateStewards.php is hacky
as phan-taint-check doesn't easily support suppressing warnings
that happen in global scope.

The rest appear to be false positives. Some of which are actual
bugs in phan-taint-check that should be fixed there. Some others
phan-taint-check can't reasonably tell what's going on and should
be suppressed, but I haven't sorted through which is which yet.

Bug: T189371
Change-Id: I82d43ef1372d31178a18e50c66dd9e2c65072a23