Update patch set 4
Patch Set 4:

This extension is a little messy when it comes to formatting details like this. Currently it contains 63 files where the initial <?php is followed by an empty line, and 38 files where this is not the case.

I believe the formatting standard you are referring to is the same I have seen a lot of times in MediaWiki core, a lot other extensions, and even a few times in this codebase here (which I'm intentionally not touching in this patch):

<?php
/**
 * File level doc
 * @file
 */

namespace …;

use …;

/**
 * Class level doc
 */
class … {
}

The relevant point here is that there are two types of comment blocks: per file, marked as such with @file, and per class. The class level comment should be connected to the class. I believe this is not technically needed for the documentation generators we use, but it makes it easier to understand for a human looking at the code.

If you go up from the class level documentation, there is a "use" section, a "namespace", and the initial "<?php". For all of this we follow the convention to separated them with newlines, as far as I know. I'm not aware of an exception for this convention, so I don't see why I shouldn't be allowed to apply it here.

Patch-set: 4