Update patch set 4
Patch Set 4:

@license and @author tags are very normal parts of the class-level doc blocks I maintained so far. If a doc block looks like it is a class-level doc block, I assume it is. Why should I assume otherwise?

Having two doc blocks for a normal file that contains a single class is very uncommon in the codebases I maintained so far. If a class does have two doc blocks, the first is a copy of the license deed, while the second describes the class.

An other example for file-level doc blocks are files that do not contain a class.

But neither nor is the case in the files I'm touching in this patch.

Patch-set: 4