Reword some messages to provide more accurate scope
The 'hidden' status does not hide the account from all public lists but
just from Special:GlobalUsers. This is misleading for people thinking
that the account is removed from public view when only the 'suppress'
option can do that.

Similarly, ending the usage of the 'oversighted' word in favor of
'suppressed' which is more accurate as for the function the term refers
to. No messages keys changed for this one yet however, as it'd imply
much more work for little gain, although this should be eventually done.

Bug: T192957
Change-Id: I479c7f212e23c6146281191440266b4be8446403