Silence iconv() warnings in CentralAuthUser
If the password has characters that can't be converted to WINDOWS-1252,
iconv() will warn and return false. We really don't care about its
warning.

I also threw in a check for false before passing the result to
$password->equals(). That shouldn't matter much, because the only way
that could return true if passed false is if the user's password is the
empty string, which means the user has bigger problems than the fact
that some password with non-WINDOWS-1252 characters will also
accidentally match.

Bug: T170971
Change-Id: If22362e73b7559f6bc94729c8da1a63067486ccb