Update patch set 1
Patch Set 1: Code-Review+2

LGTM, although I found this in the php manual that may help:

 https://secure.php.net/manual/en/function.iconv.php

 > If you append the string //IGNORE, characters that cannot be represented in the target charset are silently discarded. Otherwise, E_NOTICE is generated and the function will return FALSE.

So it seems we either discard the whole thing and get a notice with false, or we discard individual chars and get false without notice.

I guess if there was a way to let errors discard the whole thing (as now), but also not do a notice that would be nice, but if I understand correctly, that is not possible.

The main reason I was looking at this is because it would help catch an E_NOTICE for other reasons, e.g. misspelled variable name or some other error.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2