Update patch set 12
Patch Set 12: Code-Review-1

This seems rather fragile. Do we have a precedent for this kind of behaviour elsewhere in WMF-deployed MediaWiki?

Afaik we usually do one of three things:

* Attribute the action to a system user that we mark as bot properly (addGroup: bot), possibly mentioning the actor in an edit summary.
* Let the action naturally for the acting user. This means it's only marked as bots when performed by actual bots.
* Agree locally on the wiki to utilise the 'flood' right when performing the action until done.

None of these is great, but that's how we've done it so far. Is there strong reason to justify introducing a fourth way?

If is it, I propose to then build a proper way in core for doing this, if it is that important :)

For example, RecentChange::newLogEntry() could have a parameter for deciding whether to mark it as rc_bot=1, which could be passed down from MovePage::move().

Patch-set: 12
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1