Update patch set 12
Patch Set 12:

RE: the last point, it's not about permission management. The method in question already supports a query parameter for the same reason.

If we don't want to wontfix the task, then we are acknowledging this isn't a permission question, but a state / flag question. It is then possible to mark edits directly as "bot" in a way that bypasses permission management, as long as it is only exposed internally to code such as here.

But, I think that would still be confusing as the user doesn't have that right. It'd be better to use a generic system user (doesn't need to be a dedicated one, we have plenty of ones already, which have the benefit of being slightly more likely to be recognised as such). Alternatively, perhaps we should instead allow a way to explicitly supress RC entries (not just marking as bot, which is a rather poor and confusing proxy for it).

Patch-set: 12
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>