Update patch set 1
Patch Set 1:

As I understand it, there are no (and can't feasibly be) IDEs that would interpret a Gruntfile to extract its settings and feed them to an ESLint plugin for inline warnings and auto fix opportunities. This requires configuring ESLint in a way ESLint can discover. I haven't seen such an IDE, but maybe we misunderstood each other.

Invoking eslint directly with pulling in additional packages that need maintenance but don't add value, seems fine by me. I've already done that in a few repos.

Our entry point has always been 'npm test' or 'eslint/jshint/whatever', and never 'grunt' (in part so that we control the version and don't require awkward node_modules/.bin invocation and/or global installations), so our entry point will stay forward/backward compatible.

I'd be for that.

As for who made the decision, on the task most spoke in favour (except you), and in two separate FSG meetings we also agreed on .eslintignore (including this Wednesday), with this time the follow-up action item from Ed that we should document it, which I assigned to myself.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>