Add blocking details for partial blocks to Special:CentralAuth.
The Special:CentralAuth page does not give users details on the nature of
the blocks a user has. To give a better picture, the user should be shown the
details of the block (where possible).

Moved the block parameter formatting to a new method:
SpecialCentralAuth::formatBlockParams().

Moved CentralAuthUser::localUserData() to a new CentralAuthUserManager service
as UserManager::getLocalUserData().

Bug: T200938
Change-Id: Ia6ab39ef10a423e9709ae626f333765de30cdc9e