Update patch set 1
Patch Set 1: Code-Review+2

The logic is sound, and as far as I can tell, the change to the code is not going to break anything.

I'm not entirely sure it's sufficient, but it's going in the right direction.

Patch-set: 1
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2