Update patch set 34
Patch Set 34:

> Patch Set 34:
> 
> I'm not very comfortable with adding yet another parallel blocking system instead of "just" adding another behaviour action to the existing blocks system (which would be complicated as it'd need some build-out in core, but it's the "right" answer).

Note user can request renaming at _any_ wiki, not just metawiki. As such, we'd need a way how to block accounts globally, which is T17294 (sitting here since 2008). I've no idea when T17294 will be ready, but it's not likely it will be completed in foreseeable feature (no one is working on that). 

Do we really want to block this on a feature request that's sitting in Phabricator for more than ten years?

Theoretically, we can restrict the requesting special page to only metawiki, but isn't this just simpler than doing this blocking change?

Patch-set: 34
Reviewer: Urbanecm <2945@e9e9afe9-4712-486d-8885-f54b72dd1951>