Update patch set 3
Patch Set 3:

Hm, so basically a static variable in the job class, populated by the job and read by the hook. And its elements could be the IDs of the users performing a rename. Just a question: how many jobs will be executed in the same PHP process? Which is to say: should I make the registry an array, or do I only need to account for a single rename operation, and thus it could simply hold a single user's data?

Patch-set: 3
Reviewer: Daimona Eaytoy <5450@e9e9afe9-4712-486d-8885-f54b72dd1951>