Update patch set 3
Patch Set 3:

> I meant like this:
 > 
 > 
 > class LocalPageMoveJob.php {
 > public static $weIsMovinInAbout = false;
 > 
 > public function run() {
 > …
 > …
 > self::$weIsMovinInAbout = true;
 > try {
 > moveIt();
 > } finally {
 > self::$weIsMovinInAbout = false;
 > }
 > }
 > 
 > Then look for that var in the CA/AF hook.

Yeah, got it. My question was whether we're fine with a boolean property, or we need a list of users performing a rename. Which, in turn, depends on how many jobs are executed in the same PHP process.

More specifically: the LocalRenameUserJob enqueues a list of LocalPageMoveJob [1], all for the same renamer. What I don't know is how we distribute these between PHP processes. Is there one single process which runs LocalRenameUserJob and its children? Or is the former in a process, and the children are in different processes? And in the first case, will the process also run other LocalRenameUserJob for other renames, and/or completely unrelated jobs?


 > One could make it stricter by making it an array with the userID in it.

Yes, that can be done regardless of how many renames are handled in the same process. But first, I'd like to understand exactly how the jobs are processed.

 > And even more strict by making the hook "consume" the array, so that its only valid for 
 > once. :)

This is also possible, yes. As long as we remember to set it before each move(), since a single rename will involve multiple move()'s.

[1] - https://gerrit.wikimedia.org/g/mediawiki/extensions/CentralAuth/+/a50eca09ef274267aafc96e7be002acd3b235d0e/includes/LocalRenameJob/LocalRenameUserJob.php#173

Patch-set: 3
Reviewer: Daimona Eaytoy <5450@e9e9afe9-4712-486d-8885-f54b72dd1951>