Add to existing blocks if a blocked user is centrally hidden
The GetUserBlock hook handler sets a system block with a
suppression if a user's global account is suppressed (hidden).

Before this patch, if a user already has a local block with a
suppression, the handler does nothing; otherwise (if the user's
global account is suppressed) the handler overrides any blocks
they may have with the suppressing system block.

After this patch, the handler does the following if a user's
global account is suppressed:
* if the user is locally blocked and suppressed, do nothing
* if the user is not blocked, add the suppressing system block
* if the user is blocked but not suppressed, make a new
  composite block containing the existing blocks plus the
  suppressing system block

Bug: T228950
Change-Id: I665b1329bbf5ff61a9208d4e78bc6a9060988e5a