Update patch set 1
Patch Set 1: Code-Review-1

With this change, the account is no longer hidden when an admin chooses "Account is hidden from the global users list" in Special:CentralAuth.

If we wanted to allow a user to be hidden but not blocked, we'd need to reintroduce something like the UserIsHidden hook, and set the mHideName flag directly on the User, without setting a block.

I don't think we should do that - the fact that a user cannot now be hidden without a block is not a regression, but an intended consequence of deprecating the UserIsHidden hook (explained in T228948).

Patch-set: 1
Reviewer: Tchanders <2849@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1