Update patch set 5
Patch Set 5:

Looking back at the debate here from 2017, I wish I had known about Message::plaintextParam() at the time (it existed since 2014, I was just unaware of it). That seems to be the solution to the issue where the message creator doesn't know whether to escape or not since it doesn't know whether the eventual use will be text or HTML, and the message user doesn't know which parameters might need escaping (and probably shouldn't be messing with them in the first place).

Would you be willing to use that to wrap the username?

Patch-set: 5
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>