Update patch set 5
Patch Set 5:

> I'm not sure this does what this patch would need. To me it sounds more like it does the opposite.

It ensures that the username will be output as text, not parsed as wikitext. If that's not what the patch needs, what does it need?

> Also, the discussion (as I understood it) was about the Status object, which doesn't provide such methods.

The Status object is a container for messages. Either directly, or more commonly as the message key and message parameters to be used in creating a Message later.

Patch-set: 5
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>