Don't call deprecated Title::getUserCaseDBKey()
Unfortunately, the functionality was actually being used here as a hack
to normalize group names (T40883). If we just switched to getDBkey(),
group names would be uppercased (if $wgCapitalLinks is true). All this
is horrible, but I saw no good way around it without risking making
existing groups not work somewhere.  To make the hack continue working
without getUserCaseDBKey(), I instead prefixed the group name with 'A/'
before making a title, so the first letter isn't the first anymore and
won't be uppercased. This will presumably change some group names'
validity in corner cases, but hopefully it won't matter.

Bug: T202094
Change-Id: I4b8ceb8a7f4624d6a3763aca6df41bf1a0d7293f