Stop passing in 'byText' option for system and composite blocks
'byText' is used to give information about the block creator in block
error messages, but is not helpful for system and composite blocks,
which are not created by users. The information is no longer used in
error messages for these types of blocks and the option will be
obsolete after T227005.

Depends-On: Ie2aa00cfec5e166460bcaeddba3c601fe0627c13
Change-Id: I855943857c828a8607c7ba6337ed09731a86595c