Update patch set 3
Patch Set 3:

This LGTM from a code quality and API usability perspective, but I'm not familiar with the CA code itself to know whether this safe and will perform well at standard batch sizes. If this is a cheap lookup, then repeating it 500 times may be fine.

If not, then this has to be declined and/or blocked on adding batching to the internal logic because repeating it synchronously in a single thread on the server would be undesirable for prod stability (repeated and/or concurrent queries would be preferred in that case).

I've CC-ed Brad here and CPT on the task for their input, which should be waited for before merging this.

Patch-set: 3
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>