Reorder local user creation in CentralAuthTestUser
Hooks from User class may populate caches with non existing global user
GlobalPreferences is calling from hook UserSaveOptions and is using
CentralIdLookup, but that is already changed to CentralAuth one and
creating CentralAuthUser objects, but the user is not in the database.
All calls after that using the stale cache object.

Bug: T235783
Change-Id: I3c0f71e151ed6738b660379b08a25cffdef56733