Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

According to https://codesearch.wmflabs.org/search/?q=GlobalRenameBlacklist%27 there is currently only a single (which makes sense, because it's global) blacklist at https://meta.wikimedia.org/wiki/Global_rename_blacklist. As of now it contains a single entry that includes the ^…$ anchors.

I believe it does not make much of a difference if the ^…$ is hard-coded or not. If it is not (as of now), most of the entries in the blacklist need to contain the two ^…$ anchors. If the anchors *are* hard-coded (as this patch suggests), any pattern in the blacklist that is supposed to match any substring must be changed to start and/or end with .*….*. It's really just style and personal preference.

It seems it makes sense to pick the style that makes it easier to maintain the blacklist. Unfortunately the commit message does not say anything about this.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1