Update patch set 1
Patch Set 1: -Code-Review

> It seems that mostly the blacklist would be used for individual users […]

I'm not sure if this is true. When I check https://en.wikipedia.org/wiki/MediaWiki:Titleblacklist, a lot of the existing entries start with .*. It seems this one is not used for individual users. However, this might be very much different for the global blacklist here.

I agree that making these blacklists behave the same is a really good idea.

My -1 was only for the incomplete commit message. Can you update it, please, and say a little more about the motivation for this change? Also, I feel the commit message needs to say since when this feature even exists, and how we are going to update existing blacklists after the change. Thank you!

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0