Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Please do not upload more patches like this. They don't really do anything good but wasting peoples time (including your time). There is nothing wrong with assertSame. It does the exact same as assertTrue. There is no string reason to prefer one over the other, other than personal preference. And that's the issue: It's not ok to enforce one personal preference on a massive amount of essentially random codebases without knowing if the owners of these codebases agree. We do changes like this only if we reached an agreement and have a PHPCS rule in place to enforce it. This is not the case here, and unlikely to happen.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1