Update patch set 1
Patch Set 1:

> Seems sane to gather more info for this authevent.

Just to be clear, this doesn't gather more info for this authevent, so much as disable it (Default logging level for logstash is info not debug). The event triggers anytime it is checked if a user is logged in. Which is a lot since users are checked if they are logged in on a different wiki pretty constantly.

Patch-set: 1
Reviewer: Brian Wolff <99@e9e9afe9-4712-486d-8885-f54b72dd1951>