Update patch set 1
Abandoned

This patch was disputed and not updated in 4 years. The code this patch touches changed a lot in the meantime. It's most probably easier to start a new patch. This one here is still linked from the Phabricator ticket, and can still be used for inspiration, if needed. Or please feel free to reopen this patch in case you want to continue working on it.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon