CentralAuthApiTokenSessionProvider: use injected HookContainer.
New tests for this class fail if OAuth extension is installed,
breaking OAuth CI. The reason is that we mock the User which is
passed into one ApiCheckCanExecute hook, which OAuth also implements.
OAuth tries to call unmocked methods on the User making failing the test.

Instead of using global hook container, the session provider needs
to use hook container injected into it.

The APIGetAllowedParams hook is actually static, so moved it to regular
static hook registration via extension.json

Change-Id: Icfbf78a0852e61b4f03092445d09f15e84b4d9f1