Remove use of internal JavaScriptMinifier::minify() in SpecialCentralAutoLogin
This internal library should not be called directly by non-core MW
code. Use the ResourceLoader::filter() entry point instead.

Aside from abstracting which library is used and how it is called,
it also offers a caching ability.

Note that this lowers the TTL for the local-server APCu cache from
INDEFINITE to 1_DAY, which is I think is still more than long enough
to not be of any concern.

Change-Id: Id3c6dcc0b952a9efe34b3a9ed88b716101a51f87