Move conditional SpecialUsersWhoWillBeRenamed to relevant hook
This doesn't need to happen on every request. There is a relevant
hook for this that can be used.

An alternative way is to override isListed() in the
SpecialUsersWhoWillBeRenamed class, and let the execute() method
render a disabled message instead. But for now we can keep the
registration conditional as before, using the standard hook.

Bug: T180192
Change-Id: I370a4c1310b081cdc034ff0eecb804d6174e6fc3