SECURITY: GlobalRename: Avoid DoS/infinite loop in suggested username feature
If a username is already at/near the max length, trying to generate a
suggested username on Special:GlobalRenameRequest will trigger an infinite
loop because no username can satisfy the requirements.

We can avoid the infinite loop by setting a counter and giving up after 5
tries. Also add a new help message variant for when we couldn't generate a
suggested username.

Bug: T260865
Change-Id: I97d8b3236b5abed8ba9a9c4d3ab5050c2e782c22