Split database access to separate service
The utility service does a bunch of unrelated things and is in a fragile
position as it's loaded so early in the request process. This attempts
to help that by moving database-related bits into its own service.

getCentralReplicaDB() is used in SecurePoll, so that's left behind in
UtilityService for backwards compability. Otherwise moved all
UtilityService uses to the new service and dropped the old methods.

Bug: T288084
Change-Id: Ie99ac49447fb506ffa5758565848561767ba6163