Add onGetLogTypesOnUser Hook in order to treat 'gblrights' better
When doing a global rename entries in the 'Global rights log' are not
updated. This is due to the Renameuser Extension only updating logs when
their log_type is being returned from SpecialLog::getLogTypesOnUser().
Extensions can use the onGetLogTypesOnUser Hook to add more log types
to this list.

It was mentioned on the respective task, that this might not be possible
in CentralAuth because the 'gblrights' log also contains changes to the
global groups themself and to wiki sets. But that isn't quite the case
because usages of SpecialLog::getLogTypesOnUser() do also check the
value of log_namespace, which only for changes to the global user groups
of a user matches NS_USER and for all other changes is -1. So only
changes of user groups changes are going to be updated when a global
rename is performed.

Bug: T85023
Change-Id: Ifca477c9629fba6cd7ec421314d9099fedaf435a