Inject services into Hook Handlers
To prevent potential performance problems, hooks that require services
are factored out into separate hook handlers so that not all services
always have to be injected.

The SecurePoll related hook is factored into its own handler to
allow easy conversion to use interfaces for the hook once these are
introduced in the SecurePoll extension.

AbuseFilterHooks and GlobalRenameHooks are also being renamed to
AbuseFilterHookHandler and GlobalRenameHookHandler, as these are more
appropriate names for what they are doing.

Injection of a config to replace the usages of globals is going to be
done in another Follow-up patch in order to prevent this patch getting
to big.

Bug: T270846
Change-Id: I903e2043b5894dc3d4c41a07a89c8326b5d9ffae