Reenable autoblocks for CentralAuth-issued suppression blocks
This reverts commit 24bb2ed19f634ae233cedb77b113db65cdfdd889.

CentralAuth allows users with the centralauth-supress permission
to globally suppress users, e.g. due to them containing PII in
their username. Global suppressions work by placing a local block
with 'hideName' => true on each attached local account and setting
gu_hidden_level to 2 for that global user.

The big problem was that MW doesn't really support cross-wiki
blocking. It only allows passing a custom db connection to
DatabaseBlockStore::insertBlock(), which is then used to insert the
block to the database of the foreign wiki.

This means cross-wiki blocks cannot be updated or deleted. In
addition, no autoblocks could be performed cross-wiki.

But CentralAuth did, or at least tried to do so. But it actually
inserted the autoblocks into the local database and not into the
foreign one.

Note that this was not a regression, but a bug which was present
for a long time.

Secondly, when an autoblock is performed, the CheckUser extension
goes through the list of IPs used by the user and adds the last one
to the list of IPs to be autoblocked. Again, the wiki is being
ignored. This means that the CheckUser data from the local wiki is
being used.

This resulted in global suppressions throwing fatals, due to core
starting to check whether the wiki which the blocker belongs to
matches the wiki of the db connection, which it didn't.

As a security measure, autoblocking was disabled for CentralAuth
issued cross-wiki blocks in 24bb2ed19f634ae. This means de facto
that all cross-wiki autoblocks on wmf production have been disabled.

As a first step for properly fixing this we made blocks aware of
which wiki they belong to in I8ae8133f7e232c.

This now allowed the wiki to be respected when performing autoblocks,
which was implemented in I0d42c743ec188b20.

This also allows the CheckUser extension to respect the wiki, which
was implemented in Icdf9e4f37c3bfa5.

Finally we in Ia90e3578f7c103c40 we made sure that the check for
already present blocks is performed on the correct wiki.

So now autoblocks are inserted into the correct wiki, eliminating
database connection mismatches.

Therefore it is now possible to enable autoblocks on global
suppressions.

Side note: There was also the problem that with cross-wiki blocks
with restrictions, the restrictions were inserted into the local
database instead of the foreign database. This was partly fixed with
Ia342030c31710f4, the remaining bits are going to be fixed by
I1051beedfd67c4e. However, this did not affect CentralAuth anyway,
since the global suppression blocks have no restrictions.

Bug: T281972
Bug: T284873
Depends-On: I0d42c743ec188b20b636c8ec577d6cf02a7f1059
Depends-On: Icdf9e4f37c3bfa568c712bebceccf9cc46db2ec8
Depends-On: Ia90e3578f7c103c40e88f3b68ddcba4a13734d80
Depends-On: I2254e41d02ceaadff9b84e75127ca848bc920d08
Change-Id: I6165a557968e98928f1724c5554b043c371f516a