CentralAuthUserUsingDatabaseTest: don't construct CAUser
Caching in CentralAuth user depends on there being just one
instance instantiated for every user, and the constructor
is properly documented to not call directly.

What happened with recent test failures is that a mediawiki
core change 98878c08ba7f37a71c4b765f30986c755f941f85 stated
to call UserSaveSettings hook when user options were saved,
which is restoring older behaviour. That forced an instance
of CAUser to be instantiated and cached from the hook, and
that's not the same instance as the one we were running tests
on. This caused the in-memory caches between the two instances
to go out of whack.

Bug: T294265
Change-Id: I9d681baeca0df4808335e7bececfd114cdad2f0e