Drop support for CentralAuthLockedCanEdit
This is not used on Wikimedia wikis and goes fully against the idea of
locks (fully disabling an account). Let's just fully drop support for
it. Partial restrictions on global accounts should be implemented
properly (global *b*locks, likely in the GlobalBlocking extension with
CentralIdLookup support instead of being CA specific) and not like this.

Change-Id: Ic6039245fd45b3c76b942b2f399e7f1fb3e3dea0