Drop support for global user merges
Global user merges are currently not supported due to various issues,
and the UserMerge extension is unmaintained and waiting for
undeployment[0]. To partially unblock that, this patch drops all code
required for global user merges (which have been disabled on Wikimedia
wikis anyways for years).

The user merge log formatter is kept, as it is required to not break the
existing log entries. Similarly the hook handler for the DeleteAccount
hook (from UserMerge) is kept, as policy still technically requires that
all extensions support UserMerge while it is deployed on Wikimedia
wikis.

[0]: https://phabricator.wikimedia.org/T216089

Bug: T216089
Change-Id: I3860d3d644e58ced703a466aaea3433d8c0eb9fc