Introduce GlobalRenameRequestStore
This patch adds a new servive, GlobalRenameRequestStore, that will be
responsible for loading and storing GlobalRenameRequest objects. (For
now, only storing is implemented. Loading will be done in a follow-up
patch.) This is to make testing easier.

Also explicitely set the default value for GlobalRenameRequest::$deleted
field to avoid trying to write NULLs to the non-nullable database column
when updating a request that was created from scratch instead of loaded
from the database.

Change-Id: Ic919f706f2421d11354defea318dc07bd7b95831