GlobalRenameUser: Use JobQueueGroupFactory
Updates the current callable parameter to use the new
JobQueueGroupFactory class. There is one use in WikimediaMaintenance
that this will technically break, but since it is a rarely-used
maintenance script I will upload a patch[0] that can be simply merged
directly after this one so that we don't need to write temporary patches
to maintain backwards compat.

[0]: I2e3ff2a2364c98c6a8b93667d77d62ce2c294cc7

Bug: T298806
Change-Id: I87374fa317e5e0f2183ff00afbb825feec75ad22