More consistent read-only mode handling
I was surprised to discover that CentralAuthDatabaseManager::
isReadOnly() does not check $wgCentralAuthReadOnly, so callers were
not protected against subsequent exceptions from getCentralDB(). Also,
the reason was generally wrong: either incorrectly propagated from
LBFactory or thrown away. $wgCentralAuthReadOnly is documented as
being "like $wgReadOnly", which is a string, not boolean.

So:

* Make isReadOnly() check for all kinds of read-only mode.
* Remove getReadOnlyReason(). Instead add assertNotReadOnly() which
  knows how to throw an ErrorPageError which correctly displays the
  error. If the local wiki is read-only but the central database is
  not, it is incorrect to show centralauth-readonly.

Also:

* Testing showed multiple code paths leading to an exception being
  thrown from the exception handler, due to shouldUsePrimaryDB()
  returning true. So I made it return false in read-only mode.

Change-Id: I13559cc829c7ce743b51a1dae1cd631948c59e7d