SpecialGlobalGroupMembership: do not call core hooks
Removes calls to the core UserRightsChanged hook. The only consumer of
that hook in production has been Echo, which has been ignored anyways by
the hook (the task to add support, T53541, has been open since 2013).

Passing in a CentralAuthGroupMembershipProxy violates the hook signature
and broke Echo after a refactoring, so for now just remove calls to that
hook. If there is ever a need for other extensions to react to global
group changes, we can add proper hooks in CentralAuth for that.

Bug: T307518
Change-Id: Iee0b6b99a844b1aeed844a75fcf2a4e24a235874
(cherry picked from commit df2f0841ecee7d8de7602236d7065921666b116a)