Make failed login attempts store whether the password was correct
These changes, along with a few minor copyedits to qqq.json, allow
the CentralAuth extension to store reasons for failure along side
whether the password was correct. The reasons for the failure are
stored in AuthenticationResponse::$failReasons and the intended use
is to allow any code that hooks into the authentication and receives
this AuthenticationResponse to know why the request failed. The
CheckUser extension will use this information to note whether the
password was correct on login attempts to locked accounts.

Also convert the return results of CentralAuthUser::authenticate(),
CentralAuthUser::canAuthenticate() and
CentralAuthUser::authenticateWithToken() to string consts that
for backwards compatibility represent the same raw string values
previously returned.

Bug: T303192
Depends-On: I7b2d9579a518a6c02f05281b1016e31e0d086fe7
Change-Id: I379d95500f89d26882e42502a07f80e145e8359c