Remove redundant use of WRITE_SYNC
This is now a concern for documentation and implementation of the
backend rather than for individual calls. No BagOStuff implementations
other than SqlBag supported this, and afaik no callers to SqlBag
used by soon-to-be-multi-dc deployments like WMF actually use
WRITE_SYNC with SqlBag.

The flag is being removed in core with I9f59ff35bbb806.

Bug: T270225
Change-Id: I7e72c1180b7ba9e479ade62ab3dd3139d7bd5bb0