Try central autologin via redirects on login page
Browsers increasingly limit transferring cookie information
across domains via subresources by state partitioning. This
breaks CentralAuth autologin which is currently based on
subresources (AJAX, or an embedded <script> or <img> tag).

Try to circumvent limitations by doing a top-level redirect
instead. Since a top level redirect cannot be safely done in an
unknown scenario (e.g. URL fragments get lost, which might disrupt
frontend logic state), limit this to the login page. Since
autologin is already done more aggressively on the login page,
this will fit into user expectations of how autologin works.

* Add a new 'redirect' type to SpecialCentralAutoLogin's
  autologin flow (checkLoggedIn -> createSession -> validateSession
  -> setCookies), and a required "return URL" parameter to be used
  (and passed along) with this type. In redirect mode, success and
  errors are handled by redirecting back to the return URL, with
  the error message (when there is one) placed in a query parameter.
* Tokenize the "return URL" parameter using cross-wiki cache to
  prevent any open redirect vulnerability.
* Trigger a top-level redirect into this flow on SpecialUserlogin
  (using the SpecialPageBeforeExecute hook). This follows the
  logic in CentralAuthHooks::getCentralautologinJsData() and
  ext.centralauth.centralautologin.js, except for the return*
  parameters as those are already contained in the return URL
  parameter.
* Use a 'CentralAuthAnonTopLevel' cookie to remember a failed
  attempt to avoid unnecessary retries, as the redirect chain can
  add significant delay to the loading of the login page. This is
  similar to the existing 'CentralAuthAnon' localStorage flag used
  by the subresource-based autologin.
* Use a query parameter on return to prevent an infinite loop even
  if we fail to set the cookie somehow.
* Fix comments in various hooks that implied they were involved
  with redirects happening in SpecialCentralAutoLogin (while
  actually they are involved in SpecialCentralLogin).

In terms of security and caching considerations, other than the
open redirect issue, we can rely on the redirect mode's similarity
to previous modes. For open redirect prevention, the cache approach
is used because CentralAuth already requires a shared cache.
Note also that the final redirect back to the login page (which is
the only significant difference from previous modes) isn't used to
pass any sensitive data, so even if gets hijacked somehow (e.g. the
attacker manages to confuse the server about the domain of the
Special:Userlogin request), the damage is limited to the attacker
learning whether the user is currently logged in + being able to
do an open redirect.

From an UX point of view, this mode is more fragile than previous
ones (where errors were invisible to the user): if there is e.g.
an internal error in some step of the process, the user will
find themselves either on the central login wiki or
Special:CentralAutoLogin (neither of which they are familiar with,
probably), and URL state (such as a returnTo parameter) might get
lost as browsers don't typically create a history entry for a
redirect origin. In the worst case, this might prevent use of the
login page entirely; we avoid this by making sure that the
CentralAuthAnonTopLevel cookie is set right at the beginning,
before any redirection happens.

Bug: T326281
Change-Id: Ie26f60c7b55c506149cf55ed10964439fb681cad